+++ src/stream/src/executor/project/project_scalar.rs
